requirements.txt
django==3.0.1
psycopg2==2.8.4
djangorestframework==3.11.0
django_extensions==2.2.5
pyjwt==1.7.1
requests==2.22.0
sendgrid==6.1.0
django-cors-headers==3.2.1
coverage
